requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fonttools==4.53.0
h11==0.14.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
requests==2.32.3
seaborn==0.13.2
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
